+++ src/include/distributed/metadata_utility.h
@@ -66,14 +66,14 @@ typedef struct ShardInterval
 	Datum minValue;     /* a shard's typed min value datum */
 	Datum maxValue;     /* a shard's typed max value datum */
 	uint64 shardId;
-	uint64 shardGroupId;
+	int64 shardGroupId;
 	int shardIndex;
 } ShardInterval;
 
 
 typedef struct Shardgroup
 {
-	uint64 shardgroupId;
+	int64 shardgroupId;
 	uint32 colocationId;
 	Datum minShardValue; /* a shard's typed min value datum */
 	Datum maxShardValue; /* a shard's typed max value datum */
@@ -315,11 +315,11 @@ extern StringInfo GenerateSizeQueryOnMultiplePlacements(List *shardIntervalList,
 extern List * RemoveCoordinatorPlacementIfNotSingleNode(List *placementList);
 
 /* Function declarations to modify shard and shard placement data */
-extern void InsertShardGroupRow(uint64 shardGroupId, uint32 colocationId,
+extern void InsertShardGroupRow(int64 shardGroupId, uint32 colocationId,
 								text *shardMinValue, text *shardMaxValue);
 extern void InsertShardRow(Oid relationId, uint64 shardId, char storageType,
 						   text *shardMinValue, text *shardMaxValue,
-						   uint64 *shardGroupId);
+						   int64 *shardGroupId);
 extern void DeleteShardRow(uint64 shardId);
 extern uint64 InsertShardPlacementRow(uint64 shardId, uint64 placementId,
 									  uint64 shardLength, int32 groupId);
@@ -332,7 +332,7 @@ extern void UpdateDistributionColumnGlobally(Oid relationId, char distributionMe
 extern void UpdateDistributionColumn(Oid relationId, char distributionMethod,
 									 Var *distributionColumn, int colocationId);
 extern void DeletePartitionRow(Oid distributedRelationId);
-extern void DeleteShardgroupRow(uint32 shardgroupId);
+extern void DeleteShardgroupRow(int64 shardgroupId);
 extern void DeleteShardRow(uint64 shardId);
 extern void UpdatePlacementGroupId(uint64 placementId, int groupId);
 extern void DeleteShardPlacementRow(uint64 placementId);
